make a notificationcenter adapter
